requirements.txt
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.7.0
bitarray==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
cytoolz==1.0.0
distro==1.9.0
elevenlabs==1.13.3
eth-abi==5.1.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.0
packaging==24.2
parsimonious==0.10.0
pluggy==1.5.0
poly-eip712-structs==0.0.1
py-clob-client==0.18.0
py-order-utils==0.3.2
pycryptodome==3.21.0
pydantic==2.10.3
pydantic-core==2.27.1
pygame==2.6.1
pytest==8.3.4
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
sniffio==1.3.1
toolz==1.0.0
typing-extensions==4.12.2
urllib3==2.2.3
websockets==14.1
